requirements.txt
django
djangorestframework
gunicorn
django-heroku
factory_boy
pytest
pytest-django